Unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1504722 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/math3/random/RandomGeneratorAbstractTest.java
@@ -136,7 +136,7 @@ public abstract class RandomGeneratorAbstractTest extends RandomDataGeneratorTes
                 (double) ((binUpperBounds[0] + 1) * smallSampleSize) / (double) n;
             for (int k = 1; k < binCount; k++) {
                 expected[k] = (double) smallSampleSize *
-                (double) (binUpperBounds[k] - binUpperBounds[k - 1]) / (double) n;
+                (double) (binUpperBounds[k] - binUpperBounds[k - 1]) / n;
             }
             for (int j = 0; j < numTests; j++) {
                 Arrays.fill(observed, 0);
